💎feat: pr 리뷰 반영 (유니온 타입으로 변경, z-index 공용으로 관리, hook 분리)
