+++ src/app/_components/client/MonthlyGoal/hooks/index.ts
@@ -1,0 +1,1 @@
+export { default as useMonthlyGoal } from "./useMonthlyGoal/useMonthlyGoal";
